Replace ackFail with reset
