+++ Neo4j.Driver/Neo4j.Driver.IntegrationTests/Resources/rollback_error.script
@@ -1,0 +1,13 @@
+!: AUTO INIT
+!: AUTO RESET
+
+C: RUN "BEGIN" {}
+   PULL_ALL
+S: SUCCESS {}
+   SUCCESS {}
+C: RUN "CREATE (n {name:'Alice'}) RETURN n.name AS name" {}
+   PULL_ALL
+S: FAILURE {"code": "Neo.TransientError.General.DatabaseUnavailable", "message": "Unable to run any"}
+   IGNORED
+C: RESET
+S: SUCCESS {}
