+++ Neo4j.Driver/Neo4j.Driver.Tests/IO/MessageHandlers/AckFailureMessageHandlerTests.cs
@@ -15,7 +15,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System.Collections.Generic;
 using FluentAssertions;
 using Moq;
 using Neo4j.Driver.Internal.IO;
@@ -24,43 +23,38 @@ using Neo4j.Driver.Internal.Messaging;
 using Neo4j.Driver.V1;
 using Xunit;
 
-namespace Neo4j.Driver.Tests.IO.StructHandlers
+namespace Neo4j.Driver.Tests.IO.MessageHandlers
 {
-    public class FailureMessageHandlerTests : StructHandlerTests
+    public class AckFailureMessageHandlerTests : StructHandlerTests
     {
-        internal override IPackStreamStructHandler HandlerUnderTest => new FailureMessageHandler();
+        internal override IPackStreamStructHandler HandlerUnderTest => new AckFailureMessageHandler();
 
         [Fact]
-        public void ShouldThrowOnWrite()
+        public void ShouldThrowOnRead()
         {
             var handler = HandlerUnderTest;
 
             var ex = Record.Exception(() =>
-                handler.Write(Mock.Of<IPackStreamWriter>(), new FailureMessage("Code", "Message")));
+                handler.Read(Mock.Of<IPackStreamReader>(), PackStream.MsgAckFailure, 0));
 
             ex.Should().NotBeNull();
             ex.Should().BeOfType<ProtocolException>();
         }
 
         [Fact]
-        public void ShouldRead()
+        public void ShouldWrite()
         {
             var writerMachine = CreateWriterMachine();
             var writer = writerMachine.Writer();
 
-            writer.WriteStructHeader(1, PackStream.MsgFailure);
-            writer.WriteMapHeader(2);
-            writer.Write("code");
-            writer.Write("Neo.ClientError.Statement.SyntaxError");
-            writer.Write("message");
-            writer.Write("Invalid syntax.");
+            writer.Write(new AckFailureMessage());
 
             var readerMachine = CreateReaderMachine(writerMachine.GetOutput());
-            var value = readerMachine.Reader().Read();
+            var reader = readerMachine.Reader();
 
-            value.Should().NotBeNull();
-            value.Should().BeOfType<FailureMessage>().Which.Code.Should().Be("Neo.ClientError.Statement.SyntaxError");
-            value.Should().BeOfType<FailureMessage>().Which.Message.Should().Be("Invalid syntax.");
+            reader.PeekNextType().Should().Be(PackStream.PackType.Struct);
+            reader.ReadStructHeader().Should().Be(0);
+            reader.ReadStructSignature().Should().Be(PackStream.MsgAckFailure);
         }
     }
 }
