+++ Neo4j.Driver/Neo4j.Driver.Tests/IO/MessageHandlers/FailureMessageHandlerTests.cs
@@ -15,45 +15,52 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using System.Collections.Generic;
 using FluentAssertions;
 using Moq;
 using Neo4j.Driver.Internal.IO;
+using Neo4j.Driver.Internal.IO.MessageHandlers;
 using Neo4j.Driver.Internal.Messaging;
 using Neo4j.Driver.V1;
 using Xunit;
 
-namespace Neo4j.Driver.Tests.IO.StructHandlers
+namespace Neo4j.Driver.Tests.IO.MessageHandlers
 {
-    public class PullAllMessageHandlerTests : StructHandlerTests
+    public class FailureMessageHandlerTests : StructHandlerTests
     {
-        internal override IPackStreamStructHandler HandlerUnderTest => new PullAllMessageHandler();
+        internal override IPackStreamStructHandler HandlerUnderTest => new FailureMessageHandler();
 
         [Fact]
-        public void ShouldThrowOnRead()
+        public void ShouldThrowOnWrite()
         {
             var handler = HandlerUnderTest;
 
             var ex = Record.Exception(() =>
-                handler.Read(Mock.Of<IPackStreamReader>(), PackStream.MsgPullAll, 0));
+                handler.Write(Mock.Of<IPackStreamWriter>(), new FailureMessage("Code", "Message")));
 
             ex.Should().NotBeNull();
             ex.Should().BeOfType<ProtocolException>();
         }
 
         [Fact]
-        public void ShouldWrite()
+        public void ShouldRead()
         {
             var writerMachine = CreateWriterMachine();
             var writer = writerMachine.Writer();
 
-            writer.Write(new PullAllMessage());
+            writer.WriteStructHeader(1, PackStream.MsgFailure);
+            writer.WriteMapHeader(2);
+            writer.Write("code");
+            writer.Write("Neo.ClientError.Statement.SyntaxError");
+            writer.Write("message");
+            writer.Write("Invalid syntax.");
 
             var readerMachine = CreateReaderMachine(writerMachine.GetOutput());
-            var reader = readerMachine.Reader();
+            var value = readerMachine.Reader().Read();
 
-            reader.PeekNextType().Should().Be(PackStream.PackType.Struct);
-            reader.ReadStructHeader().Should().Be(0);
-            reader.ReadStructSignature().Should().Be(PackStream.MsgPullAll);
+            value.Should().NotBeNull();
+            value.Should().BeOfType<FailureMessage>().Which.Code.Should().Be("Neo.ClientError.Statement.SyntaxError");
+            value.Should().BeOfType<FailureMessage>().Which.Message.Should().Be("Invalid syntax.");
         }
     }
 }
