+++ Neo4j.Driver/Neo4j.Driver.Tests/IO/ValueHandlers/NodeHandlerTests.cs
@@ -21,16 +21,16 @@ using FluentAssertions;
 using Moq;
 using Neo4j.Driver.Internal.IO;
 using Neo4j.Driver.Internal.IO.ValueHandlers;
-using Neo4j.Driver.Internal.Messaging;
 using Neo4j.Driver.Internal.Types;
+using Neo4j.Driver.Tests.IO.MessageHandlers;
 using Neo4j.Driver.V1;
 using Xunit;
 
-namespace Neo4j.Driver.Tests.IO.StructHandlers
+namespace Neo4j.Driver.Tests.IO.ValueHandlers
 {
-    public class UnboundRelationshipHandlerTests : StructHandlerTests
+    public class NodeHandlerTests : StructHandlerTests
     {
-        internal override IPackStreamStructHandler HandlerUnderTest => new UnboundRelationshipHandler();
+        internal override IPackStreamStructHandler HandlerUnderTest => new NodeHandler();
 
         [Fact]
         public void ShouldThrowOnWrite()
@@ -39,7 +39,7 @@ namespace Neo4j.Driver.Tests.IO.StructHandlers
 
             var ex = Record.Exception(() =>
                 handler.Write(Mock.Of<IPackStreamWriter>(),
-                    new Relationship(0, -1, -1, "RELATES_TO", new Dictionary<string, object>())));
+                    new Node(0, new List<string> {"Label"}, new Dictionary<string, object>())));
 
             ex.Should().NotBeNull();
             ex.Should().BeOfType<ProtocolException>();
@@ -51,12 +51,12 @@ namespace Neo4j.Driver.Tests.IO.StructHandlers
             var writerMachine = CreateWriterMachine();
             var writer = writerMachine.Writer();
 
-            WriteUnboundRelationship(writer);
+            WriteNode(writer);
 
             var readerMachine = CreateReaderMachine(writerMachine.GetOutput());
             var value = readerMachine.Reader().Read();
 
-            VerifyWrittenUnboundRelationship(value);
+            VerifyWrittenNode(value);
         }
 
         [Fact]
@@ -66,7 +66,7 @@ namespace Neo4j.Driver.Tests.IO.StructHandlers
             var writer = writerMachine.Writer();
 
             writer.WriteListHeader(1);
-            WriteUnboundRelationship(writer);
+            WriteNode(writer);
 
             var readerMachine = CreateReaderMachine(writerMachine.GetOutput());
             var value = readerMachine.Reader().Read();
@@ -74,7 +74,7 @@ namespace Neo4j.Driver.Tests.IO.StructHandlers
             value.Should().NotBeNull();
             value.Should().BeAssignableTo<IList>().Which.Should().HaveCount(1);
 
-            VerifyWrittenUnboundRelationship(value.Should().BeAssignableTo<IList>().Which[0]);
+            VerifyWrittenNode(value.Should().BeAssignableTo<IList>().Which[0]);
         }
 
         [Fact]
@@ -85,7 +85,7 @@ namespace Neo4j.Driver.Tests.IO.StructHandlers
 
             writer.WriteMapHeader(1);
             writer.Write("x");
-            WriteUnboundRelationship(writer);
+            WriteNode(writer);
 
             var readerMachine = CreateReaderMachine(writerMachine.GetOutput());
             var value = readerMachine.Reader().Read();
@@ -94,36 +94,33 @@ namespace Neo4j.Driver.Tests.IO.StructHandlers
             value.Should().BeAssignableTo<IDictionary<string, object>>().Which.Should().HaveCount(1).And
                 .ContainKey("x");
 
-            VerifyWrittenUnboundRelationship(value.Should().BeAssignableTo<IDictionary>().Which["x"]);
+            VerifyWrittenNode(value.Should().BeAssignableTo<IDictionary>().Which["x"]);
         }
 
-        private static void WriteUnboundRelationship(IPackStreamWriter writer)
+        private static void WriteNode(IPackStreamWriter writer)
         {
-            writer.WriteStructHeader(3, PackStream.UnboundRelationship);
+            writer.WriteStructHeader(3, PackStream.Node);
             writer.Write(1);
-            writer.Write("RELATES_TO");
+            writer.Write(new List<string> {"Label1", "Label2"});
             writer.Write(new Dictionary<string, object>
             {
                 {"prop1", "something"},
-                {"prop2", 2.0},
-                {"prop3", false}
+                {"prop2", 15},
+                {"prop3", true}
             });
         }
 
-        private static void VerifyWrittenUnboundRelationship(object value)
+        private static void VerifyWrittenNode(object value)
         {
             value.Should().NotBeNull();
-            value.Should().BeOfType<Relationship>().Which.Id.Should().Be(1L);
-            value.Should().BeOfType<Relationship>().Which.StartNodeId.Should().Be(-1L);
-            value.Should().BeOfType<Relationship>().Which.EndNodeId.Should().Be(-1L);
-            value.Should().BeOfType<Relationship>().Which.Type.Should().Be("RELATES_TO");
-            value.Should().BeOfType<Relationship>().Which.Properties.Should().HaveCount(3).And.Contain(new[]
+            value.Should().BeOfType<Node>().Which.Id.Should().Be(1L);
+            value.Should().BeOfType<Node>().Which.Labels.Should().Equal(new[] { "Label1", "Label2" });
+            value.Should().BeOfType<Node>().Which.Properties.Should().HaveCount(3).And.Contain(new[]
             {
                 new KeyValuePair<string, object>("prop1", "something"),
-                new KeyValuePair<string, object>("prop2", 2.0),
-                new KeyValuePair<string, object>("prop3", false),
+                new KeyValuePair<string, object>("prop2", 15L),
+                new KeyValuePair<string, object>("prop3", true),
             });
         }
-
     }
 }
