+++ Neo4j.Driver/Neo4j.Driver.Tests/IO/ValueHandlers/Temporal/LocalTimeHandlerTests.cs
@@ -15,20 +15,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System.Collections;
-using System.Collections.Generic;
 using FluentAssertions;
-using FluentAssertions.Primitives;
-using Moq;
-using Neo4j.Driver.Internal;
 using Neo4j.Driver.Internal.IO;
 using Neo4j.Driver.Internal.IO.ValueHandlers;
-using Neo4j.Driver.Internal.Messaging;
-using Neo4j.Driver.Internal.Types;
+using Neo4j.Driver.Tests.IO.MessageHandlers;
 using Neo4j.Driver.V1;
 using Xunit;
 
-namespace Neo4j.Driver.Tests.IO.StructHandlers
+namespace Neo4j.Driver.Tests.IO.ValueHandlers
 {
     public class LocalTimeHandlerTests : StructHandlerTests
     {
