+++ Neo4j.Driver/Neo4j.Driver.Tests/IO/ValueHandlers/Temporal/SystemTimeSpanHandlerTests.cs
@@ -16,19 +16,15 @@
 // limitations under the License.
 
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using FluentAssertions;
-using FluentAssertions.Primitives;
-using Moq;
 using Neo4j.Driver.Internal.IO;
 using Neo4j.Driver.Internal.IO.ValueHandlers;
-using Neo4j.Driver.Internal.Messaging;
-using Neo4j.Driver.Internal.Types;
+using Neo4j.Driver.Tests.IO.MessageHandlers;
 using Neo4j.Driver.V1;
 using Xunit;
 
-namespace Neo4j.Driver.Tests.IO.StructHandlers
+namespace Neo4j.Driver.Tests.IO.ValueHandlers
 {
     public class SystemTimeSpanHandlerTests : StructHandlerTests
     {
