+++ Neo4j.Driver/Neo4j.Driver.Tests/IO/ValueHandlers/UnboundRelationshipHandlerTests.cs
@@ -21,16 +21,16 @@ using FluentAssertions;
 using Moq;
 using Neo4j.Driver.Internal.IO;
 using Neo4j.Driver.Internal.IO.ValueHandlers;
-using Neo4j.Driver.Internal.Messaging;
 using Neo4j.Driver.Internal.Types;
+using Neo4j.Driver.Tests.IO.MessageHandlers;
 using Neo4j.Driver.V1;
 using Xunit;
 
-namespace Neo4j.Driver.Tests.IO.StructHandlers
+namespace Neo4j.Driver.Tests.IO.ValueHandlers
 {
-    public class RelationshipHandlerTests : StructHandlerTests
+    public class UnboundRelationshipHandlerTests : StructHandlerTests
     {
-        internal override IPackStreamStructHandler HandlerUnderTest => new RelationshipHandler();
+        internal override IPackStreamStructHandler HandlerUnderTest => new UnboundRelationshipHandler();
 
         [Fact]
         public void ShouldThrowOnWrite()
@@ -39,7 +39,7 @@ namespace Neo4j.Driver.Tests.IO.StructHandlers
 
             var ex = Record.Exception(() =>
                 handler.Write(Mock.Of<IPackStreamWriter>(),
-                    new Relationship(0, 1, 2, "RELATES_TO", new Dictionary<string, object>())));
+                    new Relationship(0, -1, -1, "RELATES_TO", new Dictionary<string, object>())));
 
             ex.Should().NotBeNull();
             ex.Should().BeOfType<ProtocolException>();
@@ -51,12 +51,12 @@ namespace Neo4j.Driver.Tests.IO.StructHandlers
             var writerMachine = CreateWriterMachine();
             var writer = writerMachine.Writer();
 
-            WriteRelationship(writer);
+            WriteUnboundRelationship(writer);
 
             var readerMachine = CreateReaderMachine(writerMachine.GetOutput());
             var value = readerMachine.Reader().Read();
 
-            VerifyWrittenRelationship(value);
+            VerifyWrittenUnboundRelationship(value);
         }
 
         [Fact]
@@ -66,7 +66,7 @@ namespace Neo4j.Driver.Tests.IO.StructHandlers
             var writer = writerMachine.Writer();
 
             writer.WriteListHeader(1);
-            WriteRelationship(writer);
+            WriteUnboundRelationship(writer);
 
             var readerMachine = CreateReaderMachine(writerMachine.GetOutput());
             var value = readerMachine.Reader().Read();
@@ -74,7 +74,7 @@ namespace Neo4j.Driver.Tests.IO.StructHandlers
             value.Should().NotBeNull();
             value.Should().BeAssignableTo<IList>().Which.Should().HaveCount(1);
 
-            VerifyWrittenRelationship(value.Should().BeAssignableTo<IList>().Which[0]);
+            VerifyWrittenUnboundRelationship(value.Should().BeAssignableTo<IList>().Which[0]);
         }
 
         [Fact]
@@ -85,7 +85,7 @@ namespace Neo4j.Driver.Tests.IO.StructHandlers
 
             writer.WriteMapHeader(1);
             writer.Write("x");
-            WriteRelationship(writer);
+            WriteUnboundRelationship(writer);
 
             var readerMachine = CreateReaderMachine(writerMachine.GetOutput());
             var value = readerMachine.Reader().Read();
@@ -94,15 +94,13 @@ namespace Neo4j.Driver.Tests.IO.StructHandlers
             value.Should().BeAssignableTo<IDictionary<string, object>>().Which.Should().HaveCount(1).And
                 .ContainKey("x");
 
-            VerifyWrittenRelationship(value.Should().BeAssignableTo<IDictionary>().Which["x"]);
+            VerifyWrittenUnboundRelationship(value.Should().BeAssignableTo<IDictionary>().Which["x"]);
         }
 
-        private static void WriteRelationship(IPackStreamWriter writer)
+        private static void WriteUnboundRelationship(IPackStreamWriter writer)
         {
-            writer.WriteStructHeader(5, PackStream.Relationship);
+            writer.WriteStructHeader(3, PackStream.UnboundRelationship);
             writer.Write(1);
-            writer.Write(2);
-            writer.Write(3);
             writer.Write("RELATES_TO");
             writer.Write(new Dictionary<string, object>
             {
@@ -112,12 +110,12 @@ namespace Neo4j.Driver.Tests.IO.StructHandlers
             });
         }
 
-        private static void VerifyWrittenRelationship(object value)
+        private static void VerifyWrittenUnboundRelationship(object value)
         {
             value.Should().NotBeNull();
             value.Should().BeOfType<Relationship>().Which.Id.Should().Be(1L);
-            value.Should().BeOfType<Relationship>().Which.StartNodeId.Should().Be(2L);
-            value.Should().BeOfType<Relationship>().Which.EndNodeId.Should().Be(3L);
+            value.Should().BeOfType<Relationship>().Which.StartNodeId.Should().Be(-1L);
+            value.Should().BeOfType<Relationship>().Which.EndNodeId.Should().Be(-1L);
             value.Should().BeOfType<Relationship>().Which.Type.Should().Be("RELATES_TO");
             value.Should().BeOfType<Relationship>().Which.Properties.Should().HaveCount(3).And.Contain(new[]
             {
@@ -126,5 +124,6 @@ namespace Neo4j.Driver.Tests.IO.StructHandlers
                 new KeyValuePair<string, object>("prop3", false),
             });
         }
+
     }
 }
