+++ Neo4j.Driver/Neo4j.Driver/Internal/Connector/IConnection.cs
@@ -46,11 +46,10 @@ namespace Neo4j.Driver.Internal.Connector
         // Enqueue a run message, and a pull_all message if pullAll=true, otherwise a discard_all message 
         void Run(string statement, IDictionary<string, object> parameters = null, IMessageResponseCollector resultBuilder = null, bool pullAll = true);
 
+//        void Enqueue(IRequestMessage message1, IMessageResponseCollector responseCollector, IRequestMessage message2 = null);
+
         // Enqueue a reset message
         void Reset();
-
-        // Enqueue a ackFailure message
-        void AckFailure();
 
         /// <summary>
         /// Close and release related resources
